feat: editor support with EOS_DISABLE
Moved scripts required for plugin editor to a separate assembly to support editor UI when EOS_DISABLE flag is set
