+++ Assets/Plugins/Android/Essential.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 37957d39bd5bec14086706861b958691
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
