+++ Assets/Plugins/Android/Essential/com.playeveryware.eos.essential.asmref
@@ -1,0 +1,3 @@
+{
+    "reference": "GUID:3a63500f0eef43a438c0553491f7c5da"
+}
