+++ Assets/Plugins/Android/Essential/com.playeveryware.eos.essential.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d81a98d24f3442645b6598c0d1206b2a
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
