+++ Assets/Plugins/Essential.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1b240b3ec93bfb14dbe6dad2119ec777
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
