+++ Assets/Plugins/Essential/EOSConfig.cs
@@ -20,7 +20,9 @@
 * SOFTWARE.
 */
 
+#if !EOS_DISABLE
 using Epic.OnlineServices.Platform;
+#endif
 using System;
 using System.Collections.Generic;
 using UnityEngine;
@@ -120,6 +122,7 @@ namespace PlayEveryWare.EpicOnlineServices
             return (T)Enum.ToObject(typeof(T), value);
         }
 
+#if !EOS_DISABLE
         //-------------------------------------------------------------------------
         public static Epic.OnlineServices.IntegratedPlatform.IntegratedPlatformManagementFlags flagsAsIntegratedPlatformManagementFlags(List<string> flags)
         {
@@ -205,6 +208,13 @@ namespace PlayEveryWare.EpicOnlineServices
         }
 
         //-------------------------------------------------------------------------
+        public PlatformFlags platformOptionsFlagsAsPlatformFlags()
+        {
+            return EOSConfig.platformOptionsFlagsAsPlatformFlags(platformOptionsFlags);
+        }
+#endif
+
+        //-------------------------------------------------------------------------
         /// <summary>
         /// Creates a shallow copy of the current <c>EOSConfig</c>
         /// </summary>
@@ -229,13 +239,6 @@ namespace PlayEveryWare.EpicOnlineServices
                 && EmptyPredicates.IsEmptyOrNull(repeatButtonDelayForOverlay)
                 && EmptyPredicates.IsEmptyOrNull(initialButtonDelayForOverlay)
                 ;
-        }
-
-
-        //-------------------------------------------------------------------------
-        public PlatformFlags platformOptionsFlagsAsPlatformFlags()
-        {
-            return EOSConfig.platformOptionsFlagsAsPlatformFlags(platformOptionsFlags);
         }
 
         //-------------------------------------------------------------------------
