+++ Assets/Plugins/Essential/EOSConfig.cs.meta
