+++ Assets/Plugins/Essential/EOSPackageInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71398501e454017428d9ebcc92ddcdaf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
