+++ Assets/Plugins/Essential/EOSSteamConfig.cs
