+++ Assets/Plugins/Essential/EOSSteamConfig.cs.meta
