+++ Assets/Plugins/Essential/com.playeveryware.eos.essential.asmdef
@@ -1,12 +1,8 @@
 {
-    "name": "com.playeveryware.eos.samples",
+    "name": "com.playeveryware.eos.essential",
     "rootNamespace": "",
     "references": [
-        "GUID:2ced8f3cc58f63843bc26c7591a35628",
-        "GUID:be1f6e9efffcfce41aeb5c2f4c27cf75",
-        "GUID:073a697712d497849830fad10877c9bf",
-        "GUID:1491147abca9d7d4bb7105af628b223e",
-        "GUID:3a63500f0eef43a438c0553491f7c5da"
+        "GUID:be1f6e9efffcfce41aeb5c2f4c27cf75"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
@@ -15,7 +11,6 @@
     "precompiledReferences": [],
     "autoReferenced": true,
     "defineConstraints": [
-        "!EOS_DISABLE",
         "!EOS_DISABLE_FULL"
     ],
     "versionDefines": [],
