+++ Assets/Plugins/Essential/com.playeveryware.eos.essential.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3a63500f0eef43a438c0553491f7c5da
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
