+++ Assets/Plugins/Linux/Editor/EOSOnPostprocessBuild_Linux.cs
@@ -217,7 +217,7 @@ public class EOSOnPostprocessBuild_Linux:  IPostprocessBuildWithReport
             return eosConfig;
         }
 
-        string configFilePath = Path.Combine(Application.streamingAssetsPath, "EOS", EOSManager.ConfigFileName);
+        string configFilePath = Path.Combine(Application.streamingAssetsPath, "EOS", EOSPackageInfo.ConfigFileName);
         var configDataAsString = File.ReadAllText(configFilePath);
         var configData = JsonUtility.FromJson<EOSConfig>(configDataAsString);
         eosConfig = configData;
