+++ Assets/Plugins/Linux/Essential.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fdbfd2835cbd93544bd47f903de2ae62
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
