+++ Assets/Plugins/Linux/Essential/EOSLinuxConfig.cs.meta
