+++ Assets/Plugins/Linux/Essential/com.playeveryware.eos.essential.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8b988f4ea5d9ff14282394a11d04c6cc
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
