+++ Assets/Plugins/Source/EOS_SDK/com.Epic.OnlineServices.asmdef
@@ -9,7 +9,8 @@
     "precompiledReferences": [],
     "autoReferenced": true,
     "defineConstraints": [
-        "!EOS_DISABLE"
+        "!EOS_DISABLE",
+        "!EOS_DISABLE_FULL"
     ],
     "versionDefines": [],
     "noEngineReferences": false
