+++ Assets/Plugins/Source/Essential.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: b9468efd2774ef44cba0986502790b42
-AssemblyDefinitionReferenceImporter:
+guid: 1aa9062ddf6b3914fa104e508f51fca2
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
