+++ Assets/Plugins/Source/Essential/ICloneableGeneric.cs
