+++ Assets/Plugins/Source/Essential/IEmpty.cs.meta
