+++ Assets/Plugins/Source/Essential/com.playeveryware.eos.essential.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b6da878b4edd20b4b9b93999ed1bd590
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
