+++ Assets/Plugins/iOS/Essential.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 50c691e959e955c49bd20c0233614984
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
