+++ Assets/Plugins/iOS/Essential/com.playeveryware.eos.essential.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7045198992aaf1748aacd3511609c3ef
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
