+++ Assets/Plugins/macOS/Essential.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a98a3897ae6acd43975e74a5af7c4b9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
