+++ Assets/Plugins/macOS/Essential/EOS_macOSConfig.cs.meta
