+++ Assets/Plugins/macOS/Essential/com.playeveryware.eos.essential.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5a19bd3e96f27da4bb10f13da9911c2a
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
