--- Assets/Scripts/Editor/com.playeveryware.eos-Editor.asmref
@@ -1,3 +1,0 @@
-{
-  "reference": "GUID:cc11d9857b82d9b458fca637d5cb33fb"
-}
