+++ Assets/Scripts/Editor/com.playeveryware.eos.samples-Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 087c7cd9f29dfe24db3e39e45b3fb2d6
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
