+++ Assets/Scripts/Networking/com.playeveryware.eos.samples.networking.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "com.playeveryware.eos.networking",
+    "name": "com.playeveryware.eos..samples.networking",
     "rootNamespace": "",
     "references": [
         "GUID:2ced8f3cc58f63843bc26c7591a35628",
@@ -15,7 +15,8 @@
     "precompiledReferences": [],
     "autoReferenced": true,
     "defineConstraints": [
-        "!EOS_DISABLE"
+        "!EOS_DISABLE",
+        "!EOS_DISABLE_FULL"
     ],
     "versionDefines": [
         {
