+++ Assets/Scripts/UI/UIEOSVersion.cs
@@ -30,7 +30,7 @@ namespace PlayEveryWare.EpicOnlineServices.Samples
         private void Start()
         {
             var textComp = GetComponent<Text>();
-            textComp.text = $"v{DLLHandle.GetPackageVersion()}";
+            textComp.text = $"v{EOSPackageInfo.GetPackageVersion()}";
         }
     }
 }
